Build & test update
- updated code climate
- added .env & .env setup to easily run tests locally
- fixed dockerhub setup
- made SKIP_ABS_TESTS required for clarity
- updated python version
- removed obsolete stuff
- fix handling of root URL with trailing slash
- fixed test_custom_headers test due to changes in connection
- moved test_load_files_to_workspace to a separate file

Preparation for
https://keboola.atlassian.net/browse/AIS-35
