+++ .env.template
@@ -1,0 +1,4 @@
+KBC_TEST_TOKEN=
+# URL without trailing slash
+KBC_TEST_API_URL=
+SKIP_ABS_TESTS=
