+++ tests/mocks/test_client.py
@@ -8,4 +8,9 @@ class TestClient(unittest.TestCase):
         "Token raises attribute error when updated"
         client = Client('https://example.com', 'password')
         with self.assertRaises(AttributeError):
+            # noinspection PyPropertyAccess
             client.token = 'new-password'
+
+    def test_url_trimmed(self):
+        client = Client('https://example.com/', 'password')
+        self.assertEqual(client.root_url, 'https://example.com')
